+++ lib/features/marketplace/providers/marketplace_provider.dart
@@ -1,3 +1,4 @@
+import 'package:agricola/core/utils/error_utils.dart';
 import 'package:agricola/core/providers/analytics_provider.dart';
 import 'package:agricola/core/database/daos/marketplace_local_dao.dart';
 import 'package:agricola/core/network/http_client_provider.dart';
@@ -134,7 +135,7 @@ class MarketplaceNotifier
       _ref.read(analyticsServiceProvider).logListingCreated();
       return null;
     } catch (e) {
-      return e.toString();
+      return errorKeyFromException(e);
     }
   }
 
@@ -148,7 +149,7 @@ class MarketplaceNotifier
       );
       return null;
     } catch (e) {
-      return e.toString();
+      return errorKeyFromException(e);
     }
   }
 
@@ -160,7 +161,7 @@ class MarketplaceNotifier
       state = AsyncValue.data(current.where((l) => l.id != id).toList());
       return null;
     } catch (e) {
-      return e.toString();
+      return errorKeyFromException(e);
     }
   }
 }